Keep image browser scroll on select image
